🔧 Wrench: 스텝2의 시간선택을 스텝1으로 이전
